prevent deletion of questionnaire submissions with maximum points

+++ backend/src/main/java/net/dnadas/training_portal/service/group/project/questionnaire/QuestionnaireSubmissionService.java
@@ -169,6 +169,9 @@ public class QuestionnaireSubmissionService {
       questionnaireSubmissionDao.findByGroupIdAndProjectIdAndQuestionnaireIdAndIdAndUser(
         groupId, projectId, questionnaireId, submissionId, user).orElseThrow(
         QuestionnaireSubmissionNotFoundException::new);
+    if (questionnaireSubmission.getReceivedPoints().equals(questionnaireSubmission.getMaxPoints())) {
+      throw new IllegalStateException();
+    }
     questionnaireSubmissionDao.delete(questionnaireSubmission);
   }
